assignment1/requirements.txt
flask
gitpython
pyyaml